requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
